Update deprecated FFMPEG variables methods/members
This commit upgrades the deprecated methods/members to work with the
newest FFMPEG package (package as of March 4th 2015, hash 7da7d26).

This is done rather blindly, just replacing the given function by their
newer counterpart. No test was run to make sure this actually worked.

To period for which FFMPEG was not supported was determined by taking
the oldest and newest deprecation warning. This gives us a unsupported
period of:
 - Ostream: From 55.28.1 to 56.5.2
 - Istream: From 54.4.1 to 56.6.1